TDD on TrainerController: mockito spy-failure -tests still fail-
The now implemented function TrainerController.checkVocableOnGivenPhrase
should make all Tests pass. But this is not the case. I couldn't explain
this behavior so by having a look at the code coverage, I figured out
the following:

The true-branch of the if-statement in checkVocableOnGivenPhrase is
never been executed. This makes the test
'testCheckVocableOnGivenPhraseWhenCorrectPhrase' fail. The branch not
getting executed is caused by the statement
'correctVocable.equals(vocableToCheck)' always returning as false.
Having a look at the code coverage of the autogenerated equals-function
of Vocable.java shows that the if-statement 'if (getClass() !=
obj.getClass())' returns true and makes the equals-function return
false.
This is caused by the fact, that the Object ‚correctVocable’ is being
spied on by Mockito.
As stated in the javadoc-file of the spy-method, not an object of the
class Vocable gets created, but a spied-on copy from another
mockito-class while the spy still calls the real (not stubbed) methods
of the object of the class Vocable.

„[Spy] Creates a spy of the real object. The spy calls real methods
unless they are stubbed.“

This is a problem, since the getClass()-functions in the equals-function
will never return the same Class.
Even worse is the fact, that the other test
‚testCheckVocableOnGivenPhraseWhenIncorrectPhrase‘ now always passes
even with correct phrases, which makes it now a false positive test!

Modifying the if-statement from ‚vocableToCheck.equals(correctVocable)‘
to ‚correctVocable.equals(vocableToCheck)‘, so turning it around,
doesn’t help. Also spying on the other Vocable-object ‚vocableToCheck‘
doesn’t help either.

A possible solution would be to not spy on the object an verify the
incrementation of the counter of the Vocable.object by calling
‚getCorrTries()‘. As mentioned in your book, this solution isn’t as fine
as verifying, that the incrementation has been done by the appropriate
function ‚incCorrTries‘.

Another solution could be by keeping reference to the real
Vocable-object:

Vocable correctVocable = new Vocable(PHRASE, TRANSLATION);
Vocable spiedCorrectVocable = spy(correctVocable);
when(spiedCorrectVocable.equals(vocableToCheck)).thenReturn(correctVocable.equals(vocableToCheck));

Unfortunately the following failure by Mockito shows up:
„Also, this error might show up because:
1. you stub either of: final/private/equals()/hashCode() methods.
   Those methods *cannot* be stubbed/verified.
   Mocking methods declared on non-public parent classes is not
supported.“

The solution I will go for is the one to create a new method in the
Vocable-class ‚Vocable compareTo(Vocable v)‘.
This solution isn’t my preferred one, because now I have a
self-implemented, so not autogenerated function in my model class (in
addition to the increment-methods), which means, that my model class
contains some kind of logic, which will need to be tested by unit tests.